refactor: simplification pass over poster feature (#299)
